requirements.txt
Django~=5.1
psycopg2-binary==2.9.9
django-countries==7.6.1
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
python-dotenv~=1.0.1
django-cors-headers==4.4.0
djoser~=2.2.3